+++ CHANGELOG.md
@@ -26,6 +26,7 @@ This file contains the list of changes made to pyjoulescope_ui.
 2026 Jun 14
 
 * Added test automation support.
+* Added record for duration feature #326
 * Made the configuration portable across users and machines by substituting
   home-directory path prefixes with neutral tokens on save #312.
 * Added "Preferred prefix" to the Value & Multimeter widgets #335